requirements.txt
beautifulsoup4==4.10.0
bs4==0.0.1
Django==4.0.1
django-environ==0.8.1
djangorestframework==3.13.1
psycopg2-binary==2.9.3
python-dotenv==0.19.2
redis==4.1.0
requests==2.27.1
